refactor(rent): move rent controller DTOs into dto/ directory (#1576)
RentController already validated its request bodies with
class-validator-decorated DTOs, but they lived inline in the
controller file instead of rent/dto/ like every other module with
DTOs (bookings, agreements, payments, stellar, ...). Extracts
CalculateLateFeeDto, CalculateProratedRentDto and CreateRemindersDto
into rent/dto/ with no behavioral change, matching the repo's
established convention.

All GET endpoints in this controller take only a :id path param,
already guarded by ParseUUIDPipe.
