+++ backend/src/modules/rent/rent.controller.ts
@@ -9,123 +9,14 @@ import {
   HttpStatus,
   ParseUUIDPipe,
 } from '@nestjs/common';
-import {
-  IsNumber,
-  IsOptional,
-  IsDateString,
-  IsString,
-  IsEmail,
-  Min,
-} from 'class-validator';
-import { Type } from 'class-transformer';
-import {
-  ApiTags,
-  ApiOperation,
-  ApiResponse,
-  ApiProperty,
-  ApiPropertyOptional,
-} from '@nestjs/swagger';
+import { ApiTags, ApiOperation, ApiResponse } from '@nestjs/swagger';
 import { RentService } from './rent.service';
 import { RentReminderService } from './rent-reminder.service';
-
-// ─── DTOs ────────────────────────────────────────────────────────────────────
-
-export class CalculateLateFeeDto {
-  @ApiProperty({
-    example: 1500.0,
-    description: 'Monthly rent amount',
-    minimum: 0,
-  })
-  @IsNumber()
-  @Min(0)
-  @Type(() => Number)
-  monthlyRent: number;
-
-  @ApiProperty({
-    example: 10,
-    description: 'Number of days the payment is late',
-    minimum: 0,
-  })
-  @IsNumber()
-  @Min(0)
-  @Type(() => Number)
-  daysLate: number;
-
-  @ApiPropertyOptional({
-    example: 5,
-    description: 'Grace period in days before late fees apply',
-  })
-  @IsOptional()
-  @IsNumber()
-  @Min(0)
-  @Type(() => Number)
-  gracePeriodDays?: number;
-
-  @ApiPropertyOptional({
-    example: 0.05,
-    description: 'Daily late fee rate as a decimal (e.g., 0.05 = 5% per day)',
-  })
-  @IsOptional()
-  @IsNumber()
-  @Min(0)
-  @Type(() => Number)
-  lateFeeRate?: number;
-}
-
-export class CalculateProratedRentDto {
-  @ApiProperty({
-    example: 1500.0,
-    description: 'Full monthly rent amount',
-    minimum: 0,
-  })
-  @IsNumber()
-  @Min(0)
-  @Type(() => Number)
-  monthlyRent: number;
-
-  @ApiProperty({
-    example: '2026-06-15',
-    description: 'Move-in date for proration calculation (ISO 8601)',
-  })
-  @IsDateString()
-  moveInDate: string;
-}
-
-export class CreateRemindersDto {
-  @ApiProperty({
-    example: '123e4567-e89b-12d3-a456-426614174000',
-    description: 'UUID of the rent agreement',
-  })
-  @IsString()
-  agreementId: string;
-
-  @ApiProperty({
-    example: '123e4567-e89b-12d3-a456-426614174000',
-    description: 'UUID of the tenant',
-  })
-  @IsString()
-  tenantId: string;
-
-  @ApiProperty({
-    example: 'tenant@example.com',
-    description: 'Tenant email address for reminder delivery',
-  })
-  @IsEmail()
-  tenantEmail: string;
-
-  @ApiProperty({
-    example: '2026-07-01',
-    description: 'Rent due date (ISO 8601)',
-  })
-  @IsDateString()
-  dueDate: string;
-
-  @ApiProperty({ example: 1500.0, description: 'Rent amount due', minimum: 0 })
-  @IsNumber()
-  @Min(0)
-  @Type(() => Number)
-  amount: number;
-}
+import {
+  CalculateLateFeeDto,
+  CalculateProratedRentDto,
+  CreateRemindersDto,
+} from './dto';
 
 // ─── Controller ──────────────────────────────────────────────────────────────
